requirements.txt
fastapi==0.110.0
pymongo==4.6.1
pandas==2.2.1
pydantic==2.6.3
requests==2.31.0
uvicorn==0.27.1
gunicorn==21.2.0
numpy==1.26.4
dnspython==2.6.1
python-dateutil==2.9.0
python-dotenv==1.0.1
